% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\docType{class}
\name{knapsack-class}
\alias{knapsack-class}
\alias{knapsack}
\title{knapsack problem}
\description{
The package contains three different functions for solving what is called the knapsack problem.
The knapsack problem is a discrete optimization problem where we have a knapsack that can take a
limited weight 'W' and we want to fill this knapsack with a number of items i = 1, ..., n, each with a
weight w_i and a value v_i. The goal is to find the knapsack with the largest value of the elements added
to the knapsack.
}

