% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAB6C.R
\docType{package}
\name{LAB6C}
\alias{LAB6C}
\alias{LAB6C-package}
\title{LAB6C}
\description{
The package contains three different functions for solving what is called the knapsack problem.
}
\author{
Zijie Feng

Andreas C.charitos

Muhammad Daud
}
